% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CenteredPartition.R
\name{CenteredPartition}
\alias{CenteredPartition}
\title{Probabilities for the Centered Partition Process}
\usage{
CenteredPartition(anchor, shrinkage, baseline, useVI = TRUE, a = 1)
}
\arguments{
\item{anchor}{An integer vector giving the anchor partition
(a.k.a., center partition, location partition).}

\item{shrinkage}{A scalar giving the value of the shrinkage parameter.}

\item{baseline}{An object of class \code{PartitionDistribution}
representing a partition distribution.  Currently, only
\code{\link{UniformPartition}}, \code{\link{JensenLiuPartition}} and
\code{\link{CRPPartition}} are supported.}

\item{useVI}{Should the distance between a particular partition and the
anchor partition be measured using the variation of information
(\code{TRUE}) or using Binder loss (\code{FALSE})?}

\item{a}{A nonnegative scalar influencing the relative cost of placing two items
in separate clusters when in truth they belong to the same cluster.  This
defaults to \eqn{1}, meaning equal costs.}
}
\value{
A numeric vector giving either probabilities or log probabilities for
the supplied partitions.
}
\description{
This function specifies the centered partition distribution for given an
anchor partition, shrinkage, concentration, and discount
parameters.
}
\examples{
CenteredPartition(anchor=c(1,1,1,2,2), shrinkage=3,
                  baseline=CRPPartition(nItems=5, concentration=1.5, discount=0.1))

CenteredPartition(anchor=c(1,1,1,2,2), shrinkage=3,
                  baseline=UniformPartition(nItems=5))
}
