% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRPPartition.R
\name{CRPPartition}
\alias{CRPPartition}
\title{Probabilities for the Chinese Restaurant Process (CRP) Partition Distribution}
\usage{
CRPPartition(nItems, concentration, discount = 0)
}
\arguments{
\item{nItems}{The number of items in each partition, as an integer greater
than or equal to \eqn{1}.}

\item{concentration}{The concentration parameter, as a numeric value
greater than \eqn{-1*discount}.}

\item{discount}{The discount parameter, as a numeric value in \eqn{[0,1)}.}
}
\value{
An object of class \code{PartitionDistribution} representing this
partition distribution.
}
\description{
This function specifies the Chinese restaurant process (CRP) partition
distribution for given concentration and discount parameters.
}
\examples{
CRPPartition(nItems=5, concentration=1.5, discount=0.1)
}
