% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FixedPartition.R
\name{FixedPartition}
\alias{FixedPartition}
\title{Probabilities for the Fixed Partition Distribution}
\usage{
FixedPartition(anchor)
}
\arguments{
\item{anchor}{An integer vector giving the anchor partition
(a.k.a., center partition, location partition).}
}
\value{
An object of class \code{PartitionDistribution} representing this
partition distribution.
}
\description{
This function specifies the fixed partition distribution for a given anchor
partition.  This is a point-mass distribution at the anchor partition.
}
\examples{
FixedPartition(anchor=c(1,1,1,2,2))
}
