% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JensenLiuPartition.R
\name{JensenLiuPartition}
\alias{JensenLiuPartition}
\title{Probabilities for the Jensen Liu Partition Distribution}
\usage{
JensenLiuPartition(concentration, permutation)
}
\arguments{
\item{concentration}{The concentration parameter, as a numeric value
greater than \eqn{-1*discount}.}

\item{permutation}{A vector containing the integers \eqn{1, 2, \ldots, n}
giving the order in which items are allocated to the partition.}
}
\value{
An object of class \code{PartitionDistribution} representing this
partition distribution.
}
\description{
This function specifies the Jensen Liu (2008) partition distribution for given
concentration and permutation parameters.
}
\examples{
JensenLiuPartition(concentration=0.5, permutation=c(1,5,4,2,3))
}
