% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkDistrPtr.R
\name{mkDistrPtr}
\alias{mkDistrPtr}
\title{Make a Pointer to Partitional Distribution Parameters}
\usage{
mkDistrPtr(distr, excluded = NULL, included = NULL)
}
\arguments{
\item{distr}{An object of class \sQuote{PartitionDistribution}.}

\item{excluded}{A character vector of explicitly excluded partition distributions.}

\item{included}{A character vector of explicitly included partition distributions.}
}
\value{
A pointer
}
\description{
Users should not call this function.  This is an internal function exported
for the sake of developers of packages depending on this package.
}
