% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prPartition.R
\name{prPartition}
\alias{prPartition}
\title{Compute the (Log of) the Probability of a Partition}
\usage{
prPartition(distr, partition, log = TRUE)
}
\arguments{
\item{distr}{A specification of the partition distribution, i.e., an object
of class \code{PartitionDistribution} as returned by, for example, a
function such as \code{\link{CRPPartition}}.}

\item{partition}{A matrix of integers giving cluster labels on the rows. For
partition \eqn{k}, items \eqn{i} and \eqn{j} are in the same cluster if and
only if \code{partition[k,i] == partition[k,j]}.}

\item{log}{A logical indicating whether the probability (\code{FALSE}) or its
natural logarithm (\code{TRUE}) is desired.}
}
\value{
A numeric vector giving either probabilities or log probabilities for
the supplied partitions.
}
\description{
This function computes (the log of) the probability of the supplied
partitions for the given partition distribution.
}
\examples{
concentration <- 1.0
discount <- 0.1
nSamples <- 3

distr <- CRPPartition(nItems=5, concentration=concentration, discount=discount)
x <- samplePartition(distr, nSamples, nCores=1)
prPartition(distr, x)

anchor <- c(1,1,1,2,2)
permutation <- c(1,5,4,2,3)
n_items <- length(permutation)

distr <- ShrinkagePartition(anchor=anchor, shrinkage=c(0,0,0,0.3,0.3),
             permutation=permutation, grit=0.2,
             CRPPartition(nItems=n_items, concentration=concentration, discount=discount))
x <- samplePartition(distr, nSamples, nCores=1)
prPartition(distr, x)
}
\seealso{
\code{\link{CRPPartition}}, \code{\link{ShrinkagePartition}},
\code{\link{LocationScalePartition}}, \code{\link{CenteredPartition}},
\code{\link{samplePartition}}
}
