% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplePartition.R
\name{samplePartition}
\alias{samplePartition}
\title{Sample From a Partition Distribution}
\usage{
samplePartition(
  distr,
  nSamples,
  randomizePermutation = FALSE,
  randomizeShrinkage = c("fixed", "common", "cluster", "idiosyncratic")[1],
  randomizeGrit = FALSE,
  shrinkage_shape = 5,
  shrinkage_rate = 1,
  grit_shape1 = 1,
  grit_shape2 = 1,
  nCores = 0
)
}
\arguments{
\item{distr}{A specification of the partition distribution, i.e., an object
of class \code{PartitionDistribution} as returned by, for example, a
function such as \code{\link{CRPPartition}}.}

\item{nSamples}{An integer giving the number of partitions to sample.}

\item{randomizePermutation}{Should the permutation be uniformly randomly
sampled for each partition?}

\item{randomizeShrinkage}{Should the shrinkage be random for each
sample?  Specifically, the shrinkage is the same for every
observations and sampled from a gamma distribution with parameters
\code{shrinkage_shape} and \code{shrinakge_rate}.}

\item{randomizeGrit}{Should the grit of the \code{\link{ShrinkagePartition}}
be sampled from a beta distribution for each partition? This is ignored
for other distributions.}

\item{shrinkage_shape}{The shape parameter of the gamma distribution for
randomizing the shrinkage.}

\item{shrinkage_rate}{The rate parameter of the gamma distribution for
randomizing the shrinkage.}

\item{grit_shape1}{The first parameter of the beta distribution for
randomizing the grit.}

\item{grit_shape2}{The first parameter of the beta distribution for
randomizing the grit.}

\item{nCores}{The number of CPU cores to use. A value of zero indicates to
use all cores on the system.}
}
\value{
An integer matrix containing a partition in each row using cluster
label notation.
}
\description{
This function samples from a partition distribution.
}
\details{
Note that the centered partition distribution \code{\link{CenteredPartition}}
uses MCMC to sample.
}
\examples{
concentration <- 1.0
discount <- 0.1
nSamples <- 3

distr <- CRPPartition(nItems=5, concentration=concentration, discount=discount)
x <- samplePartition(distr, nSamples, nCores=1)
prPartition(distr, x)

anchor <- c(1,1,1,2,2)
permutation <- c(1,5,4,2,3)
n_items <- length(permutation)

distr <- ShrinkagePartition(anchor=anchor, shrinkage=c(0,0,0,0.3,0.3),
             permutation=permutation, grit=0.2,
             CRPPartition(nItems=n_items, concentration=concentration, discount=discount))
x <- samplePartition(distr, nSamples, nCores=1)
prPartition(distr, x)
}
\seealso{
\code{\link{CRPPartition}}, \code{\link{ShrinkagePartition}},
\code{\link{LocationScalePartition}}, \code{\link{CenteredPartition}},
\code{\link{prPartition}}
}
