% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShrinkagePartition.R
\name{ShrinkagePartition}
\alias{ShrinkagePartition}
\title{Probabilities for the Shrinkage Partition Distribution}
\usage{
ShrinkagePartition(anchor, shrinkage, permutation, grit, baseline)
}
\arguments{
\item{anchor}{An integer vector giving the anchor partition
(a.k.a., center partition, location partition).}

\item{shrinkage}{A numeric vector of length equal to the length of
\code{anchor} (i.e., the number of items) giving the shrinkage
probability for each item. This can also be a scalar, in which case that
value is used for each item.}

\item{permutation}{A vector containing the integers \eqn{1, 2, \ldots, n}
giving the order in which items are allocated to the partition.}

\item{grit}{A numeric value controlling the amount of clustering, with small
values encouraging few clusters and large values encouraging more clusters.
Values between 0 and 1 (exclusive) ensure that, as shrinakge goes to
infinity, the partition distribution concentrates on the anchor partition.}

\item{baseline}{An object of class \code{PartitionDistribution}
representing a partition distribution.  Currently, only
\code{\link{UniformPartition}}, \code{\link{JensenLiuPartition}} and
\code{\link{CRPPartition}} are supported.}
}
\value{
An object of class \code{PartitionDistribution} representing this
partition distribution.
}
\description{
This function specifies the shrinkage partition distribution given an anchor
partition, shrinkage, permutation and a baseline distribution.
}
\examples{
ShrinkagePartition(anchor=c(1,1,1,2,2), shrinkage=c(10,10,10,3,3),
                   permutation=c(1,5,4,2,3), grit=0.1,
                   baseline=CRPPartition(nItems=5, concentration=1.5, discount=0.1))

}
