% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UniformPartition.R
\name{UniformPartition}
\alias{UniformPartition}
\title{Probabilities for the Uniform Partition Distribution}
\usage{
UniformPartition(nItems)
}
\arguments{
\item{nItems}{An integer giving the number of items in each partition.}
}
\value{
An object of class \code{PartitionDistribution} representing this
partition distribution.
}
\description{
This function specifies the uniform partition distribution.
}
\examples{
UniformPartition(nItems=5)
}
