% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplePartition.R
\name{samplePartition}
\alias{samplePartition}
\title{Sample From a Partition Distribution}
\usage{
samplePartition(
  distr,
  nSamples,
  randomizePermutation = FALSE,
  randomizeShrinkage = c("fixed", "common", "cluster", "idiosyncratic")[1],
  max = 4,
  shape1 = 1,
  shape2 = 1,
  nCores = 0
)
}
\arguments{
\item{distr}{A specification of the partition distribution, i.e., an object
of class \code{PartitionDistribution} as returned by, for example, a
function such as \code{\link{CRPPartition}}.}

\item{nSamples}{An integer giving the number of partitions to sample.}

\item{randomizePermutation}{Should the permutation be uniformly randomly
sampled for each partition?}

\item{randomizeShrinkage}{Should the shrinkage be random for each
permutation?  Specifically, the shrinkage is the same for every
observations and sampled from a beta distribution with parameters
\code{shape1} and \code{shape2}, as scaled to have a maximum value of
\code{max} (instead of 1).}

\item{max}{The scalar giving the maximum value of the shrinkage.}

\item{shape1}{The first shape parameter of the beta distribution for
randomizing the shrinkage.}

\item{shape2}{The second shape parameter of the beta distribution for
randomizing the shrinkage.}

\item{nCores}{The number of CPU cores to use. A value of zero indicates to
use all cores on the system.}
}
\value{
An integer matrix containing a partition in each row using cluster
label notation.
}
\description{
This function samples from a partition distribution.
}
\details{
Note that the centered partition distribution \code{\link{CenteredPartition}}
is not supported.
}
\examples{
concentration <- 1.0
discount <- 0.1
nSamples <- 3

distr <- CRPPartition(nItems=5, concentration=concentration, discount=discount)
x <- samplePartition(distr, nSamples, nCores=1)
prPartition(distr, x)

anchor <- c(1,1,1,2,2)
permutation <- c(1,5,4,2,3)
n_items <- length(permutation)

distr <- ShrinkagePartition(anchor=anchor, shrinkage=c(0,0,0,0.3,0.3),
             permutation=permutation, grit=0.2,
             CRPPartition(nItems=n_items, concentration=concentration, discount=discount))
x <- samplePartition(distr, nSamples, nCores=1)
prPartition(distr, x)
}
\seealso{
\code{\link{CRPPartition}}, \code{\link{ShrinkagePartition}},
\code{\link{LocationScalePartition}}, \code{\link{CenteredPartition}},
\code{\link{prPartition}}
}
